Update cups-filters to 1.21.1
Summary:
Bug fix release, especially to fix a crash bug on foomatic-rip introduced by a set of Coverity Scan fixes.
- foomatic-rip: Fixed segmentation fault caused by wrong Coverity Scan issue fix.
- Build system: Require QPDF 8.1.0 or later as it is needed by bannertopdf.

Signed-off-by: Pierre-Yves <pyu@riseup.net>

Test Plan:
Our package was already patched for the first issue and the 2nd one is not a problem since QPDF 8.2.1 is already available on Solus.
However since it's quiet today, I made the update.

Reviewers: #triage_team, DataDrake

Reviewed By: #triage_team, DataDrake

Subscribers: DataDrake

Differential Revision: https://dev.solus-project.com/D3706